Move view holder factory classes to correct package
* Move QuestionnaireItemBarCodeReaderViewHolderFactoryInstrumentedTest class to contrib package
* Move QuestionnaireItemBarCodeReaderViewHolderFactory class to contrib package
