+++ datacapturegallery/src/main/java/com/google/android/fhir/datacapture/gallery/CustomQuestionnaireFragment.kt
@@ -17,7 +17,7 @@
 package com.google.android.fhir.datacapture.gallery
 
 import com.google.android.fhir.datacapture.QuestionnaireFragment
-import com.google.android.fhir.datacapture.contrib.QuestionnaireItemBarCodeReaderViewHolderFactory
+import com.google.android.fhir.datacapture.contrib.views.barcode.QuestionnaireItemBarCodeReaderViewHolderFactory
 
 class CustomQuestionnaireFragment : QuestionnaireFragment() {
   override fun getCustomQuestionnaireItemViewHolderFactoryMatchers():
